Fix data types
Fix a bug on 32-bit arch (#108), and clarify 32/64 bit sizes. See #110
for code-review.

Squashed these commits:

minor, change temporary buffer size

minor, change temporary buffer size.

Fix data types

+ Change data type of dataset length (number of rows)
  from int to DSLength (currently uint64_t)
+ Change data type of NumEven per zmw from int to DNALength
  (currently uint32_t)

CommandLineParser: int should be sufficient to cover num of command line args

remove a warning: GFFFile forget to assign attributes.

remove a warning: AfgBasWriter

remove a warning: StringUtils.

remove a warning: PackedDNASequence

Remove a warning from HDFData.

remove a warning from HDFAtom

remove  a warning from HDFAttributable

Remove a warning from BufferedHDFArrayImpl

remove a warning from BufferedHDF2DArrayImpl

remove a warning from HDFWriterBase

Remove a warning from HDFBaseCallsWriter

remove a warning from HDFZMWWriter

remove a warning from HDFBaxWriter

remove a warning from HDFPulseCallsWriter

Remove a warning from HDFPulseWriter

Remove a warning from DNASequence

Remove a warning from FASTQSequence

Remove warnings from SMRTSequence

Fix data type in CCSSequence.hpp

Remove a warning from FASTAReader

Remove a warning from SMRTReadUtils.

Fix data type in BaseFile.

Fix data type in PulseFile.

Remove a warning in reads/ScanData

Fix data type of HoleNumber in RegionAnnotation and RegionTable.

ALARMING, Regions/HoleNumber is of data type int32_t,
while ZMW/HoleNumber is of data type uint32_t in SPEC.
blasr_libcpp should be aware of the discrepency, and
handle Regions carefully.

Fix data type in CmpAlignment

Fix data type

Fix data type in AlnGroup, MovieInfo, RefGroup, RefInfo

Fix data types in ReadGroup, ReferenceSequence, SAMAlignment, SAMHeader, SAMKeywordValuePair

Add -Wall and -Wextra to blasr_libcpp

Define GenomeLength and TupleData

CompressedSequence: use GenomeLength instead of long

pbdata/utils: upgrade to uint64_t

utils/BitUtils: remove used code.

FASTAReader/FASTQReader: replace long with uint64

BufferedHDFArray: add support to types int32_t and int64_t.

HDFPlsReader: fix type of dataset index from int to uint64_t

ScanData: fix BaseMapToStr

Remove a warning in MappingMetrics

Remove a warning from FileOfFileName

Remove warnings from RegionUtils (num of subreads per zmw should be a small integer.

Remove warnings from ExtendedAlign.

Remove warnings from GuidedeAlign

Remove warnings from *ScoreFunction

Fix data type in DifferenceCovers.

Remove warnings from LighgweightSuffixArray.

Remove warnings from MultikeyQuicksort

Fix data type in Alignment.cpp and AlignmentMap.cpp

Remove warnings.

FilterCriteria: check for undefined path and bail out.

Fix data types and remove warnings from SAMToAlignmentCandidateAdapter

Remove warnings in FragmentCCSIterator

Remove a warning from ReaderAgglomerate.

Remove  a warning from SAMSupplementalQVList

Fix data type in BaseTuple.

Fix data type in DNATuple

subsequent change to commit 8648430

Fix data type in TupleCountTable.

Subsequent changes to commit 8648430

Fix data type in TupleMask and TupleMetrics.

Subsequent changes to commit 8648430

Update data type in AnchorParamters

Remove a warning from MatchPos

Change data type and remove a warning from WeightedInterval.

Update ScoreAnchorsImpl.hpp since TupleCountTable changed in commit d651ba9

Remove a warning from CompareSequencesPrinterImpl

Remove a warning in SAMHeaderPrinter assuming num of target sequences that we can handle in a SAM/BAM file is less than max int32_t

Remove warnings from SAMPrinter.

remove unused code in pdf.

Remove a warning in LookupAnchorDistribution

Remove a warning from  LengthHistogram.cpp

Fix data type and remove warnings from ContextSample.

Add -Wall -Wextra to alignment/makefile.

Replace std::fill with vector(size, value)

ScanData: check BaseMap

Remove a warning from CompressedSequenceImpl

use vector.assign to avoid using additional tmp array.

Remove a warning from HDFCCSReader.

+++ alignment/algorithms/alignment/GuidedAlign.hpp
@@ -29,6 +29,9 @@
 
 #define MAX_BAND_SIZE 250
 
+//FIXME: Change data type of target pos to GenomeLength in order to support
+//       > 4G genome.
+
 class GuideRow {
 public:
     int q, t;
@@ -55,7 +58,7 @@ void StoreMatrixOffsets(Guide &guide);
 
 float QVToLogPScale(char qv);
 
-void QVToLogPScale(QualityValueVector<QualityValue> &qualVect, int phredVectLength, std::vector<float> &lnVect); 
+void QVToLogPScale(QualityValueVector<QualityValue> &qualVect, size_t phredVectLength, std::vector<float> &lnVect); 
 
 int AlignmentToGuide(blasr::Alignment &alignment, Guide &guide, int bandSize);
